refactor: Refactor Tuple's id to make Update support complex queries (#263)
* refactor: Refactor Tuple's id to make Update support complex queries

* chore: codefmt

+++ src/expression/mod.rs
@@ -1501,7 +1501,7 @@ mod test {
                 distinct: true,
                 kind: AggKind::Avg,
                 args: vec![ScalarExpression::Empty],
-                ty: LogicalType::Integer,
+                ty: LogicalType::Double,
             },
             Some((&transaction, &table_cache)),
             &mut reference_tables,
